Handle RPC server disconnects
Once the xmlrpc connection has been created, there will be an open
connection to the lava server, but if the connection drops, the rpc
calls will fail. The call failure was not be trapped and the exception
would terminate the entire scheduler. Now the dropped connection is
trapped and the scheduler will attempt to recreate a connection. Any
currently unexpected exceptions will now be trapped at the top level
loop and logged and the watcher should be able to continue.